update names and move to correct files
