+++ src/measurement_state.cpp
@@ -1,4 +1,4 @@
-#include "MeasurementState.h"
+#include "measurement_state.h"
 
 
 MeasurementState::MeasurementState(uint8_t pirPin, unsigned long holdDuration)
