--- src/networking_base.ino
@@ -1,11 +1,0 @@
-#include "networking_base.h"
-
-WifiClient wifi;
-EthernetClient ether;
-NetworkingBase network (&wifi, &ether) ;
-
-void setup() {
-}
-
-void loop() {
-}
